Move all env.public to a new place
