+++ apps/designer/app/routes/$.tsx
@@ -4,7 +4,7 @@ import { useLoaderData } from "@remix-run/react";
 import djb2a from "djb2a";
 import { InstanceRoot, Root } from "@webstudio-is/react-sdk";
 import { loadCanvasData } from "~/shared/db";
-import env, { type Env } from "~/env.server";
+import env, { type PublicEnv } from "~/env/env.public";
 import { sentryException } from "~/shared/sentry";
 import { Canvas } from "~/canvas";
 import { ErrorMessage } from "~/shared/error";
@@ -19,7 +19,7 @@ import type { CanvasData } from "@webstudio-is/project";
 import { customComponents } from "~/canvas/custom-components";
 import { createContext } from "~/shared/context.server";
 
-type Data = CanvasData & { env: Env; mode: BuildMode };
+type Data = CanvasData & { env: PublicEnv; mode: BuildMode };
 
 export const dynamicLinks: DynamicLinksFunction<CanvasData> = ({
   data,
